working process

+++ Assets/Scripts/Input.cs
@@ -9,27 +9,27 @@ public class Input : ScriptableObject
 
     [Header("Grid")]
     public Vector2Int n = new Vector2Int(50, 51);
-    public double lx = 1.0d;
-    public double ly = 1.0d;
+    public float lx = 1.0f;
+    public float ly = 1.0f;
 
     [Header("Time")]
-    public double tf = 10.0d;
+    public float tf = 10.0f;
     public int frame = 10;
 
     [Header("Physical Variables")]
-    public double re = 100;
-    public double density = 1.0d;
-    public double cfl = 1.0d;
-    public double fo = 0.1d;
-    public double characteristicLength = 1.0;
-    public double characteristicSpeed = 1.0;
+    public float re = 100;
+    public float density = 1.0f;
+    public float cfl = 1.0f;
+    public float fo = 0.1f;
+    public float characteristicLength = 1.0f;
+    public float characteristicSpeed = 1.0f;
     public string scheme = "CD4";
 
     [Header("Initial Condition")]
     public bool custom = false;
-    public double[] uSides = { 0.0d, 0.0d, 0.0d, 1.0d };
-    public double[] vSides = { 0.0d, 0.0d, 0.0d, 0.0d };
-    public double uIni = 0.0d;
-    public double vIni = 0.0d;
+    public float[] uSides = { 0.0f, 0.0f, 0.0f, 1.0f };
+    public float[] vSides = { 0.0f, 0.0f, 0.0f, 0.0f };
+    public float uIni = 0.0f;
+    public float vIni = 0.0f;
 
 }
